Last commit for 6.19
Imported ex_flash.c and ex_flash.h file written by Matt, however, there are a lot of external parameters that is not defined in this system which is casuing 26 errors now. So need to fix that. Further need to read Charpter 7 of the FreeRTOS Turiol guide since ex_flash file also includes Semaphores which i have no idea about; and have not configured at all in CubeMx.