Finish setting up STPM32 and USART3
Finish a test procedure for STPM32 and USART3. Turn STPM32 into block state every 3 seconds. This solves the problem of printing error for USART3 as now USART3 will not be occupied (since STPM32 is not using it anymore). Also added writing to flash function in USART3 thread, havent test it but guess it will be okay.