Fixed the problem of STM32 trigger Hardfault
Fixed the problem of STM32 trigger HardFault error, this was caused by a stack overflow from the task: SPI2. As LoRaMAC uses a lot of stack space, the original 128 byte is not enought to store that mucn data. As a result, increased the stack size from 128 to 256 in task initialization state in the beginning of FreeRTOS.c.