Solved the problem of DIO0 interrupt not triggered corrctly
FYI, solved the problem by setting the varaible "GpioIrq" to global, as it was local and two same name ones was created. So the GpioIrq[] in IRQ init was not the same one used in the callback; thus making program unable to trigger DIO IRQ correctly. It is solved now.
Next try to solve the problem of not repeatly setting the OP mode back to IDLE, still working on it.